Swapped views -> init for events
