+++ ftplugin/k8s_root.lua
@@ -1,5 +1,5 @@
 local deployment_view = require("kubectl.views.deployments")
-local event_view = require("kubectl.views.events.views")
+local event_view = require("kubectl.views.events")
 local node_view = require("kubectl.views.nodes.views")
 local service_view = require("kubectl.views.services.views")
 local secret_view = require("kubectl.views.secrets.views")
