+++ lua/kubectl/init.lua
@@ -2,7 +2,7 @@ local commands = require("kubectl.actions.commands")
 local config = require("kubectl.config")
 local pod_view = require("kubectl.views.pods")
 local deployment_view = require("kubectl.views.deployments")
-local events_view = require("kubectl.views.events.views")
+local events_view = require("kubectl.views.events")
 local nodes_view = require("kubectl.views.nodes.views")
 local secrets_view = require("kubectl.views.secrets.views")
 local services_view = require("kubectl.views.services.views")
