+++ lua/kubectl/views/events/init.lua
@@ -1,35 +1,25 @@
+local actions = require("kubectl.actions.actions")
+local definition = require("kubectl.views.events.definition")
+local ResourceBuilder = require("kubectl.resourcebuilder")
+
 local M = {}
 
-function M.processRow(rows)
-  local data = {}
-  for _, row in pairs(rows.items) do
-    local pod = {
-      namespace = row.metadata.namespace,
-      lastseen = row.lastTimestamp,
-      type = row.type,
-      reason = row.reason,
-      object = row.involvedObject.name,
-      message = row.message,
-      count = row.count,
-    }
-
-    table.insert(data, pod)
-  end
-  return data
+function M.Events()
+  ResourceBuilder:new("events", { "get", "events", "-A", "-o=json" })
+    :fetch()
+    :decodeJson()
+    :process(definition.processRow)
+    :prettyPrint(definition.getHeaders)
+    :addHints({
+      { key = "<enter>", desc = "message" },
+    }, true, true)
+    :setFilter(FILTER)
+    :display("k8s_events", "Events")
 end
 
-function M.getHeaders()
-  local headers = {
-    "NAMESPACE",
-    "LASTSEEN",
-    "TYPE",
-    "REASON",
-    "OBJECT",
-    "MESSAGE",
-    "COUNT",
-  }
-
-  return headers
+function M.ShowMessage(event)
+  local msg = event
+  actions.floating_buffer(vim.split(msg, "\n"), "event_msg", { title = "Message", syntax = "less" })
 end
 
 return M
